requirements.txt
discord
python-decouple
pyyml